[chore]채팅 리팩터링
